+++ app/src/main/java/org/megamind/mycashpoint/data/data_source/remote/dto/solde/SoldeResponse.kt
@@ -10,7 +10,7 @@ import java.math.BigDecimal
 @Serializable()
 data class SoldeResponse(
     val id: Long,
-    val operateurId: Int,
+    val operateurId: Long,
     val operateurName: String,
     val soldeType: SoldeType,
     @Serializable(with = BigDecimalSerializer::class)
